+++ src/enums/index.ts
@@ -21,3 +21,6 @@ export * from './RequestMethod';
 export * from './RouteAction';
 export * from './AuthMethod';
 export * from './SessionMode';
+export * from './StatusCode';
+export * from './ContentType';
+export * from './RequestHeader';